requirements.txt
Django==2.2.10
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.2